# fix: Improve performance of UI updates (Part II)