--- a/unity-client/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarName.cs
+++ b/unity-client/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarName.cs
@@ -33,7 +33,7 @@ public class AvatarName : MonoBehaviour
 
             for (int i = 0; i < layoutGroupRTs.Count; i++)
             {
-                Utils.ForceRebuildLayoutImmediate<LayoutGroup>(layoutGroupRTs[i]);
+                Utils.ForceRebuildLayoutImmediate(layoutGroupRTs[i]);
             }
 
             RefreshTextPosition();