--- a/unity-client/Assets/Scripts/MainScripts/DCL/Components/UI/UIImage/UIImage.cs
+++ b/unity-client/Assets/Scripts/MainScripts/DCL/Components/UI/UIImage/UIImage.cs
@@ -90,7 +90,7 @@ namespace DCL.Components
             referencesContainer.paddingLayoutGroup.padding.left = Mathf.RoundToInt(model.paddingLeft);
             referencesContainer.paddingLayoutGroup.padding.right = Mathf.RoundToInt(model.paddingRight);
 
-            LayoutRebuilder.ForceRebuildLayoutImmediate(parentRecTransform);
+            Utils.ForceRebuildLayoutImmediate<LayoutGroup>(parentRecTransform);
             return null;
         }