--- a/unity-client/Assets/Scripts/MainScripts/DCL/Components/UI/UIShape.cs
+++ b/unity-client/Assets/Scripts/MainScripts/DCL/Components/UI/UIShape.cs
@@ -202,8 +202,9 @@ namespace DCL.Components
             }
 
             uiGameObject =
-                UnityEngine.Object.Instantiate(Resources.Load(prefabPath), parentUIComponent.childHookRectTransform) as
-                    GameObject;
+                UnityEngine.Object.Instantiate(
+                    Resources.Load(prefabPath),
+                    parentUIComponent != null ? parentUIComponent.childHookRectTransform : null) as GameObject;
             referencesContainer = uiGameObject.GetComponent<T>();
 
             referencesContainer.rectTransform.SetToMaxStretch();
@@ -249,8 +250,6 @@ namespace DCL.Components
                 model.width.GetScaledValue(parentTransform.rect.width));
             referencesContainer.layoutElementRT.SetSizeWithCurrentAnchors(RectTransform.Axis.Vertical,
                 model.height.GetScaledValue(parentTransform.rect.height));
-
-            referencesContainer.layoutElementRT.ForceUpdateRectTransforms();
         }
 
         public void RefreshDCLAlignmentAndPosition(RectTransform parentTransform = null)
@@ -262,19 +261,7 @@ namespace DCL.Components
 
             referencesContainer.layoutElement.ignoreLayout = false;
             ConfigureAlignment(referencesContainer.layoutGroup);
-
-            // NOTE(Santi): It seems to be very much cheaper to execute these 4 instructions in an independet way than
-            //              execute directly the function 'LayoutRebuilder.ForceRebuildLayoutImmediate(referencesContainer.layoutGroup.transform as RectTransform)',
-            //              that theorically already contains these 4 instructions.
-            HorizontalLayoutGroup[] layoutGroups = referencesContainer.layoutGroup.GetComponentsInChildren<HorizontalLayoutGroup>();
-            for (int i = 0; i < layoutGroups.Length; i++)
-            {
-                layoutGroups[i].CalculateLayoutInputHorizontal();
-                layoutGroups[i].CalculateLayoutInputVertical();
-                layoutGroups[i].SetLayoutHorizontal();
-                layoutGroups[i].SetLayoutVertical();
-            }
-
+            Utils.ForceRebuildLayoutImmediate<HorizontalLayoutGroup>(parentTransform);
             referencesContainer.layoutElement.ignoreLayout = true;
 
             // Reposition