--- a/unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarEditorHUD/Scripts/NFTItemInfo.cs
+++ b/unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarEditorHUD/Scripts/NFTItemInfo.cs
@@ -88,8 +88,7 @@ public class NFTItemInfo : MonoBehaviour
         Utils.InverseTransformChildTraversal<LayoutGroup>((x) =>
         {
             RectTransform rt = x.transform as RectTransform;
-            rt.ForceUpdateRectTransforms();
-            LayoutRebuilder.ForceRebuildLayoutImmediate(rt);
+            Utils.ForceRebuildLayoutImmediate<LayoutGroup>(rt);
         }, transform);
 
         if (gameObject.activeInHierarchy)