--- a/unity-client/Assets/Scripts/MainScripts/Debugging/Panels/StatsPanelToggler.cs
+++ b/unity-client/Assets/Scripts/MainScripts/Debugging/Panels/StatsPanelToggler.cs
@@ -61,7 +61,7 @@ namespace DCL
             }
 
             miscController.gameObject.SetActive(!miscController.gameObject.activeSelf);
-            Utils.ForceRebuildLayoutImmediate<LayoutGroup>(transform.parent.GetComponent<RectTransform>());
+            Utils.ForceRebuildLayoutImmediate(transform.parent.GetComponent<RectTransform>());
         }
 
         private void EnableDeepProfiling()