Improve the performance related to the hiccups provoked by the UI updates (part 2).

- [x] Avoid the use of `RectTransform -> ForceUpdateRectTransforms()` when it is not needed.
- [x] Avoid the use of `TextMeshProUGUI -> ForceMeshUpdate()` when it is not needed.
- [x] Replace all `LayoutRebuilder.ForceRebuildLayoutImmediate()` calls by our custom function.
- [x] Review and refactor all  places where `LayoutRebuilder.ForceRebuildLayoutImmediate()` was being used.
- [x] Test it and measure times.

**Evolution of the improved performance**
These metrics have been taken on some places where I could reproduce always the SAME related spikes (derived from `UIShape -> RefreshAll()`) in a very specific way:

- **Genesis Plaza [0,0]** (opening the UI box clicking on the robot):
  **BEFORE ANY FIX:** ~55.30ms
  **FIX - PART 1 ([already merged PR](https://github.com/decentraland/explorer/pull/1166)):** ~18.71ms
  **FIX - PART 2 (this PR):** ~10.54ms

- **Infinity Engine [-68,-36]** (just after the scene has been loaded and `RendererState` is set to ON):
  **BEFORE ANY FIX:** ~790.30ms
  **FIX - PART 1 ([already merged PR](https://github.com/decentraland/explorer/pull/1166)):** ~228.49ms
  **FIX - PART 2 (this PR):** ~87.18ms